feat: add securitization with token for endpoints create, delate and update

+++ users/app/services/users_services.py
@@ -7,14 +7,10 @@ from app.config.logger import get_custom_logger
 from app.database import db_manager
 from app.schemas.user_schemas import UserToken, UserCredential, UserCreate
 from app.utils.authentication import create_access_token
-from app.utils.constants.error import USER_NOT_FOUND, EMAIL_ALREADY_EXISTS, BAD_CREDENTIALS
+from app.utils.constants.error import USER_NOT_FOUND, EMAIL_ALREADY_EXISTS, BAD_CREDENTIALS, DELETING_SUPERADMIN
 from app.utils.encrypt import get_password_hash, verify_password
 
 logger = get_custom_logger(logging.getLogger(__name__))
-
-
-def get_user_by_email(db: Session, user_email: str):
-    return db_manager.get_user_by_email(db, user_email)
 
 
 def create_user(db: Session, user: UserCreate):
@@ -37,6 +33,10 @@ def delete_user(db: Session, user_email: str):
     if not db_user:
         logger.error("The user with this email not exist")
         raise HTTPException(status_code=500, detail={"code": USER_NOT_FOUND})
+
+    if db_user.superadmin:
+        logger.error("The user is superadmin, cannot delete it")
+        raise HTTPException(status_code=500, detail={"code": DELETING_SUPERADMIN})
 
     return db_manager.delete_user(db, db_user)
